requirements.txt
boto3==1.26.88
botocore==1.29.88
jmespath==1.0.1
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
scipy==1.10.1
six==1.16.0
tqdm==4.65.0
urllib3==1.26.14
pyyaml
matplotlib
pytorch-lightning
transformers
wandb